+++ PlataformaVideojuegos/src/main/java/org/example/model/form/updates/GameUpdate.java
@@ -13,7 +13,7 @@ public record GameUpdate(
         String developer,
         LocalDate launchDate,
         float basePrice,
-        float currentDescount,
+        int currentDescount,
         String category,
         GameAgeClasification gameAgeClasification,
         List<String> availabeLanguages,
